# allocates less buffers when decoding multiple large protos