… examples/filter-simple/main.wasm
… examples/score-simple/main.wasm
… guest/testdata/all/main.wasm
… guest/testdata/filter/main.wasm
… guest/testdata/score/main.wasm